Rework presenters to reuse resources
