+++ app/views/content_items/worldwide_office.html.erb
@@ -1,5 +1,5 @@
 <%= render partial: "content_items/worldwide_organisation/header", locals: {
-  link_to_organisation: true,
+  worldwide_organisation: @content_item.worldwide_organisation,
 } %>
 
 <article class="govuk-grid-row">
