+++ app/views/content_items/worldwide_organisation/_contact.html.erb
@@ -1,7 +1,7 @@
 <%= content_tag(:div, class: "worldwide-organisation-contact-section", lang: local_assigns[:lang] ? lang : nil) do %>
   <% unless local_assigns[:hide_title] %>
     <%= render "govuk_publishing_components/components/heading", {
-      text: contact["title"],
+      text: contact.title,
       heading_level: 3,
       font_size: "s",
       margin_bottom: 2,
