+++ app/views/content_items/worldwide_organisation/_header.html.erb
@@ -11,12 +11,12 @@
     <div class="govuk-grid-row">
       <div class="govuk-grid-column-full worldwide-organisation-header__metadata">
         <dl>
-          <% if @content_item.world_location_links.present? %>
+          <% if worldwide_organisation&.world_location_links.present? %>
             <dt><%= t('worldwide_organisation.location') %>:</dt>
-            <dd data-module="hide-other-links"><%= @content_item.world_location_links %></dd>
+            <dd data-module="hide-other-links"><%= worldwide_organisation.world_location_links %></dd>
           <% end %>
           <dt><%= t('worldwide_organisation.part_of') %>:</dt>
-          <dd lang="en"><%= @content_item.sponsoring_organisation_links %></dd>
+          <dd lang="en"><%= worldwide_organisation&.sponsoring_organisation_links %></dd>
         </dl>
       </div>
     </div>
